refactor: localize frontend service abstractions
