+++ Sakila.Web/Abstractions/IApiClient.cs
@@ -1,7 +1,7 @@
 using FluentValidation;
-using Sakila.Contracts.Common;
+using Sakila.Web.Abstractions;
 
-namespace Sakila.Web.Common;
+namespace Sakila.Web.Abstractions;
 
 public interface IApiClient
 {
