+++ Sakila.Web/Extensions/ServiceCollectionExtensions.cs
@@ -4,9 +4,9 @@ using Sakila.Contracts.Countries.Commands;
 using Sakila.Contracts.Countries.Commands.Validators;
 using Sakila.Contracts.Languages.Commands;
 using Sakila.Contracts.Languages.Commands.Validators;
-using Sakila.Contracts.Services;
+using Sakila.Web.Abstractions;
 using Sakila.Web.Common;
-using Sakila.Web.Services;
+using Sakila.Web.Services.Implementations;
 
 namespace Sakila.Web.Extensions;
