+++ Sakila.Web/Pages/Countries/List.razor.cs
@@ -1,8 +1,7 @@
 using Microsoft.AspNetCore.Components;
-using Sakila.Contracts.Common;
+using Sakila.Web.Abstractions;
 using Sakila.Contracts.Countries.Commands;
 using Sakila.Contracts.Countries.Queries.Responses;
-using Sakila.Contracts.Services;
 
 namespace Sakila.Web.Pages.Countries;
