+++ Sakila.Web/Pages/Languages/List.razor.cs
@@ -1,8 +1,7 @@
 using Microsoft.AspNetCore.Components;
-using Sakila.Contracts.Common;
+using Sakila.Web.Abstractions;
 using Sakila.Contracts.Languages.Commands;
 using Sakila.Contracts.Languages.Queries.Responses;
-using Sakila.Contracts.Services;
 
 namespace Sakila.Web.Pages.Languages;
